move bench out of test folder, cleanup
